Weights Color & Warning Clean

+++ SkinMesh.cpp
@@ -48,10 +48,10 @@ void SkinMesh::init()
 {
   ColoredRenderable::init();
 
-  Renderable::computeNormals<VertexData_Colored>(m_vertices.data(), m_vertices.size(),
-                                                 m_indices.data(), m_indices.size(), 3, true);
+  Renderable::computeNormals<VertexData_Colored>(m_vertices.data(), (int) m_vertices.size(),
+                                                 m_indices.data(), (int) m_indices.size(), 3, true);
 
-  ColoredRenderable::initRenderable(m_vertices.data(), m_vertices.size(), m_indices.data(), m_indices.size());
+  ColoredRenderable::initRenderable(m_vertices.data(), (int) m_vertices.size(), m_indices.data(), (int) m_indices.size());
 }
 
 // ------------------------------------------------------------------------------------------------
@@ -66,7 +66,7 @@ void SkinMesh::update(UpdateInfo infos)
     verticesUpdate();
   }
 
-  ColoredRenderable::updateRenderable(GL_TRIANGLES, m_indices.size());
+  ColoredRenderable::updateRenderable(GL_TRIANGLES, (int) m_indices.size());
 }
 
 // ------------------------------------------------------------------------------------------------
@@ -102,8 +102,8 @@ void SkinMesh::setRelation(const RigPtr& rig)
 // ------------------------------------------------------------------------------------------------
 void SkinMesh::verticesUpdate()
 {
-  Renderable::computeNormals<VertexData_Colored>(m_vertices.data(), m_vertices.size(),
-                                                 m_indices.data(), m_indices.size(), 3, true);
+  Renderable::computeNormals<VertexData_Colored>(m_vertices.data(), (int) m_vertices.size(),
+                                                 m_indices.data(), (int) m_indices.size(), 3, true);
 
-  Renderable::updateVertices<VertexData_Colored>(m_vertices.data(), m_vertices.size());
+  Renderable::updateVertices<VertexData_Colored>(m_vertices.data(), (int) m_vertices.size());
 }
